refactor(semantic): check duplicate parameters in Binder of FormalParameters
